requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
colorama==0.4.6
coverage==7.6.1
Django==4.2.15
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-pytest==0.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
Faker==12.0.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
mixer==7.2.2
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.7.0
